2일차/requirements.txt
langchain
langgraph
langchain-openai
openai
langchain-community
tavily-python
faiss-cpu
pypdf
python-dotenv
pandas 
tiktoken
langchainhub